requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.3
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.7.0
